requirements.txt
anyio==4.4.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
google-auth==2.33.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.0.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
urllib3==2.2.2
